[ALPS02942047] TEEI: kernel
[Detail]
Add Copyright and cancel command etc from Microtrust vendor.

MTK-Commit-Id: 92213be4f4107d319ec65f01225bf0cb3815308d

CR-Id: ALPS02942047
Feature: Microtrust TEEI

+++ drivers/misc/mediatek/teei/V1.0/tz_driver/include/irq_register.h
@@ -1,3 +1,17 @@
+/*
+ * Copyright (c) 2015-2016 MICROTRUST Incorporated
+ * All Rights Reserved.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * version 2 as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ */
+
 #include "utdriver_macro.h"
 
 struct work_entry {
@@ -51,8 +65,8 @@ extern struct service_handler socket;
 extern struct service_handler vfs_handler;
 extern struct service_handler printer_driver;
 
-extern unsigned int forward_call_flag;
-extern unsigned int soter_error_flag;
+extern unsigned long forward_call_flag;
+extern unsigned long soter_error_flag;
 extern struct semaphore smc_lock;
 extern struct completion global_down_lock;
 extern unsigned long teei_config_flag;
